@@ -10,15 +10,18 @@
     "@mui/icons-material": "^5.2.5",
     "@mui/lab": "^5.0.0-alpha.62",
     "@mui/material": "^5.2.5",
+    "@mui/x-date-pickers": "^6.3.1",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.24.0",
+    "dayjs": "^1.11.7",
+    "dotenv": "^16.0.3",
     "i18next": "^21.6.4",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.29.1",
     "prop-types": "^15.8.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-animated-slider-2": "^1.1.5",
     "react-copy-to-clipboard": "^5.0.4",
     "react-datepicker": "^4.5.0",
