@@ -1,18 +1,20 @@
 import React, { useState, useEffect } from "react";
-import Photogrid from "./PhotoGrid";
-import axios from "axios";
-import { apiUrl } from "../../Default";
-const HomeGaleri = () => {
-  const [images, setImages] = useState([]);
-  useEffect(() => {
-    getGaleriHome();
-  }, []);
-  const getGaleriHome = () => {
-    axios.get(`${apiUrl}/galeri/home?`).then((response) => {
-      const imageData = response.data.data.map((v, i) => v.image);
-      setImages(imageData.slice(0, 6));
-    });
-  };
+import ImageList from "@mui/material/ImageList";
+import ImageListItem from "@mui/material/ImageListItem";
+const BASE_URL =
+  process.env.NODE_ENV === "production"
+    ? process.env.REACT_APP_API_URL_PROD
+    : process.env.REACT_APP_API_URL_DEV;
+
+const HomeGaleri = ({ images }) => {
+  function srcset(image, size, rows = 1, cols = 1) {
+    return {
+      src: `${image}?w=${size * cols}&h=${size * rows}&fit=crop&auto=format`,
+      srcSet: `${image}?w=${size * cols}&h=${
+        size * rows
+      }&fit=crop&auto=format&dpr=2 2x`,
+    };
+  }
   return (
     <div className="homeGaleriContainer">
       <div className="servicesTitle">
@@ -25,7 +27,27 @@ const HomeGaleri = () => {
         </div>
       </div>
       <div className="galeriGrid">
-        <Photogrid images={images} />
+        <ImageList variant="quilted" cols={4} rowHeight={300}>
+          {images &&
+            images.map((item) => (
+              <ImageListItem
+                key={item.img}
+                cols={item.cols || 1}
+                rows={item.rows || 1}
+              >
+                <img
+                  {...srcset(
+                    `${BASE_URL}/${item.file_url}`,
+                    121,
+                    item.rows,
+                    item.cols
+                  )}
+                  alt={item.title}
+                  loading="lazy"
+                />
+              </ImageListItem>
+            ))}
+        </ImageList>
       </div>
     </div>
   );
