@@ -1,13 +1,16 @@
-import React, { useState, useLayoutEffect, useEffect } from "react";
+import React, { useState, useLayoutEffect } from "react";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import EllipseOrange from "./assets/EllipseOrange.png";
 import EllipseBlue from "./assets/EllipseBlue.png";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
-import { apiUrl } from "../../Default";
 import { cardDescSlicer, cardTitleSlicer } from "../utils/funcHelper";
+
+const BASE_URL =
+  process.env.NODE_ENV === "production"
+    ? process.env.REACT_APP_API_URL_PROD
+    : process.env.REACT_APP_API_URL_DEV;
 
 function useWindowSize() {
   const [size, setSize] = useState([0, 0]);
@@ -21,21 +24,11 @@ function useWindowSize() {
   }, []);
   return size;
 }
-const PelatihanPopuler = () => {
+const PelatihanPopuler = ({ training }) => {
   const [cardIdx, setCardContentIdx] = useState(0);
   const [width, height] = useWindowSize();
-  const [popularData, setPopularData] = useState([]);
   let navigate = useNavigate();
 
-  useEffect(() => {
-    getPelatihanPopular();
-  }, []);
-
-  const getPelatihanPopular = () => {
-    axios.get(`${apiUrl}/pelatihan/popular`).then((response) => {
-      setPopularData(response.data.data);
-    });
-  };
   const settings = {
     slidesToShow: width <= 969 ? 1 : width <= 1200 ? 2 : 3,
     slidesToScroll: 1,
@@ -56,66 +49,66 @@ const PelatihanPopuler = () => {
         </div>
         <div className="pelatihanCardContainer">
           <Slider {...settings}>
-            {popularData.map((v, i) => {
-              return (
-                <div key={i}>
-                  <div
-                    className="pelatihanCardInner"
-                    onMouseEnter={() => setCardContentIdx(i)}
-                    style={{
-                      boxShadow:
-                        cardIdx === i
-                          ? "rgba(99, 99, 99, 0.2) 0px 2px 8px 0px"
-                          : "",
-                      backgroundColor: "#FFFFFF",
-                    }}
-                  >
-                    <div className="imageContainer">
-                      <div
-                        style={{
-                          width: "100%",
-                          height: "100%",
-                          backgroundImage: `url(${apiUrl}/${v.image})`,
-                          backgroundRepeat: "no-repeat",
-                          backgroundSize: "cover",
-                          backgroundPosition: "center",
-                          cursor: "pointer",
-                        }}
-                      />
-                    </div>
+            {training &&
+              training.length > 0 &&
+              training.map((v, i) => {
+                return (
+                  <div key={i}>
                     <div
-                      className="pelatihanCardTitle"
+                      className="pelatihanCardInner"
+                      onMouseEnter={() => setCardContentIdx(i)}
                       style={{
-                        color: cardIdx === i ? "#FDC232" : "#000000",
+                        boxShadow:
+                          cardIdx === i
+                            ? "rgba(99, 99, 99, 0.2) 0px 2px 8px 0px"
+                            : "",
+                        backgroundColor: "#FFFFFF",
                       }}
                     >
-                      {cardTitleSlicer(v.name)}
-                    </div>
-                    <div className="pelatihanCardDesc">
-                      {cardDescSlicer(v.descriptions)}
-                    </div>
-                    <div className="pelatihanCardBtn">
-                      <div className="cardBtnTitle">
-                        <a
-                          onClick={() =>
-                            navigate(
-                              "/info/detailPelatihan/" + v.id + "/" + v.name
-                            )
-                          }
+                      <div className="imageContainer">
+                        <div
                           style={{
-                            textDecoration: "none",
-                            color: cardIdx === i ? "#FDC232" : "#000000",
+                            width: "100%",
+                            height: "100%",
+                            backgroundImage: `url(${BASE_URL}/${v.training_image})`,
+                            backgroundRepeat: "no-repeat",
+                            backgroundSize: "cover",
+                            backgroundPosition: "center",
                             cursor: "pointer",
                           }}
-                        >
-                          Lihat Detail
-                        </a>
+                        />
+                      </div>
+                      <div
+                        className="pelatihanCardTitle"
+                        style={{
+                          color: cardIdx === i ? "#FDC232" : "#000000",
+                        }}
+                      >
+                        {cardTitleSlicer(v.name)}
+                      </div>
+                      <div className="pelatihanCardDesc">
+                        {cardDescSlicer(v.descriptions)}
+                      </div>
+                      <div className="pelatihanCardBtn">
+                        <div className="cardBtnTitle">
+                          <a
+                            onClick={() =>
+                              navigate("/info/detailPelatihan/" + v.id)
+                            }
+                            style={{
+                              textDecoration: "none",
+                              color: cardIdx === i ? "#FDC232" : "#000000",
+                              cursor: "pointer",
+                            }}
+                          >
+                            Lihat Detail
+                          </a>
+                        </div>
                       </div>
                     </div>
                   </div>
-                </div>
-              );
-            })}
+                );
+              })}
           </Slider>
         </div>
       </div>
